// src/types/skills/ElementUnion.ts
type ElementUnion =
  | "fir" // fire
  | "pas" // passive
  | "sup" // support
  | "lig" // light
  | "ice" // ice
  | "ele" // electric
  | "psy" // psychic
  | "win" // wind
  | "ail" // ailment
  | "gun" // gun
  | "dar" // curse
  | "nuk" // nuke
  | "alm" // almighty
  | "rec" // healing
  | "phy" // physical
  | "none";

export default ElementUnion;